refactor[server/utils] Move updateTruckInfo and updateEventInfo to the utils folder
Also modified require statements at the top of the files
