+++ utils/updateTruckInfo.js
